+++ App.tsx
@@ -1,16 +1,14 @@
 import {View, Text, SafeAreaView, ScrollView} from 'react-native';
 import React from 'react';
-import Sensors from './components/sensors';
+import Sensors from './src/components/sensors';
 
 const App = () => {
   return (
     <ScrollView>
       <SafeAreaView>
         <View style={{margin: 20}}>
-          <Text>
-          React-Native-Sensor
-          </Text>
-     
+          <Text>React-Native-Sensor</Text>
+
           <Sensors />
         </View>
       </SafeAreaView>
